refactor: move Cilium status types to shared module and update hook implementation
